hirewise/hirewise/hirewise/requirements.txt
Flask>=2.0
PyPDF2>=3.0
python-docx>=0.8.11
nltk>=3.5
reportlab>=4.0
gunicorn>=20.1.0
itsdangerous>=2.0
Werkzeug>=2.0
Jinja2>=3.0
python-dateutil>=2.8.2

